RESTful_Face_Web/requirement.txt
alabaster==0.7.10
anaconda-client==1.6.3
anaconda-navigator==1.6.3
anaconda-project==0.6.0
asn1crypto==0.22.0
astroid==1.4.9
astropy==2.0
Babel==2.4.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.10.1
bleach==1.5.0
bokeh==0.12.6
boto==2.47.0
Bottleneck==1.2.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cloudpickle==0.2.2
clyent==1.2.2
colorama==0.3.9
conda==4.3.22
configparser==3.5.0
contextlib2==0.5.5
coreapi==2.3.1
coreschema==0.0.4
cryptography==1.8.1
cycler==0.10.0
Cython==0.25.2
cytoolz==0.8.2
dask==0.15.0
datashape==0.5.4
decorator==4.0.11
distributed==1.17.1
Django==1.11.3
django-rest-swagger==2.1.2
djangorestframework==3.6.3
docutils==0.13.1
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==0.12.2
Flask-Cors==3.0.2
gevent==1.2.2
greenlet==0.4.12
h5py==2.7.0
HeapDict==1.0.0
html5lib==0.999
httpie==0.9.9
idna==2.5
imagesize==0.7.1
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.2.15
itsdangerous==0.24
itypes==1.1.0
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-core==4.3.0
lazy-object-proxy==1.3.1
llvmlite==0.19.0
locket==0.2.0
lxml==3.8.0
Markdown==2.6.8
MarkupSafe==0.23
matplotlib==2.0.2
mistune==0.7.4
mpmath==0.19
msgpack-python==0.4.8
multipledispatch==0.4.9
MySQL-python==1.2.5
mysqlclient==1.3.10
navigator-updater==0.1.0
nbconvert==5.2.1
nbformat==4.3.0
networkx==1.11
nltk==3.2.4
nose==1.3.7
notebook==5.0.0
numba==0.34.0
numexpr==2.6.2
numpy==1.12.1
numpydoc==0.6.0
odo==0.5.0
olefile==0.44
openapi-codec==1.3.2
openpyxl==2.4.7
packaging==16.8
pandas==0.20.2
pandocfilters==1.4.1
partd==0.3.8
pathlib2==2.2.1
patsy==0.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.2
ply==3.10
prompt-toolkit==1.0.14
psutil==5.2.2
ptyprocess==0.5.1
py==1.4.34
pycosat==0.6.2
pycparser==2.17
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.5.0
Pygments==2.2.0
pylint==1.6.4
PyMySQL==0.7.11
pyodbc==4.0.17
pyOpenSSL==17.0.0
pyparsing==2.1.4
pytest==3.1.2
python-dateutil==2.6.0
pythondialog==3.4.0
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.0
QtPy==1.2.1
requests==2.14.2
requests-toolbelt==0.8.0
rope-py3k==0.9.4.post1
scikit-image==0.13.0
scikit-learn==0.18.2
scipy==0.19.1
seaborn==0.7.1
simplegeneric==0.8.1
simplejson==3.11.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
Sphinx==1.6.2
sphinxcontrib-websupport==1.0.1
spyder==3.1.4
SQLAlchemy==1.1.11
statsmodels==0.8.0
sympy==1.1
tables==3.3.0
tblib==1.3.2
terminado==0.6
testpath==0.3
toolz==0.8.2
tornado==4.5.1
traitlets==4.3.2
unicodecsv==0.14.1
uritemplate==3.0.0
uWSGI==2.0.15
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==2.0.0
wrapt==1.10.10
xlrd==1.0.0
XlsxWriter==0.9.6
xlwt==1.2.0
zict==0.1.2
